Disable sidebar interaction entirely in viewing mode
